requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.0.11
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-extensions==3.2.3
django-ninja==1.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
Faker==26.0.0
h11==0.14.0
idna==3.7
pillow==10.3.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
whitenoise==6.8.2
